Finesse overall CSS styling
Developer note: I didn't get this project quite where I wanted to
before leaving for a month's worth of travel.  Realizing I should have
committed what I had (or at least written a commit message) before
heading out, even if the project as a whole wasn't finalized.

For the sake of continuing progress, though, I'm submitting this as is
and considering it a lesson learned.